Add starting point feature. Fixes #5

+++ src/process_input.rs
@@ -1,14 +1,14 @@
 use digits::BaseCustom;
 
-pub fn derive_min_max(range: &str) -> (i32, i32) {
+pub fn derive_min_max(range: &str) -> (usize, usize) {
   let rvals = range.split(':').collect::<Vec<&str>>();
   let mut rivals = rvals.iter();
   let min = rivals.next().unwrap();
   let max = rivals.next();
-  let min = min.parse::<i32>().unwrap();
-  let get_max = || -> i32 {
+  let min = min.parse::<usize>().unwrap();
+  let get_max = || -> usize {
     match max {
-      Some(v) => { v.parse::<i32>().unwrap() },
+      Some(v) => { v.parse::<usize>().unwrap() },
       _ => { min }
     }
   };
